requirements.txt
pyaudio
websockets
superu
python-dotenv
